perf: code quality
