+++ dashboard/src/components/shared/AstrBotConfig.vue
@@ -4,7 +4,7 @@ import { ref, computed } from 'vue'
 import ConfigItemRenderer from './ConfigItemRenderer.vue'
 import TemplateListEditor from './TemplateListEditor.vue'
 import { useI18n, useModuleI18n } from '@/i18n/composables'
-import { usePluginI18n } from '@/utils/pluginI18n'
+import { useConfigTextResolver } from '@/composables/useConfigTextResolver'
 import axios from 'axios'
 import { useToast } from '@/utils/toast'
 
@@ -40,21 +40,8 @@ const props = defineProps({
 })
 
 const { t } = useI18n()
-const { tm, getRaw } = useModuleI18n('features/config-metadata')
-const { configText } = usePluginI18n()
-
-const translateIfKey = (value) => {
-  if (!value || typeof value !== 'string') return value
-  return getRaw(value) ? tm(value) : value
-}
-
-const resolveConfigText = (path, attr, fallback) => {
-  const fallbackText = translateIfKey(fallback) || ''
-  if (!props.pluginName || !props.pluginI18n || Object.keys(props.pluginI18n).length === 0) {
-    return fallbackText
-  }
-  return configText(props.pluginI18n, path, attr, fallbackText)
-}
+const { getRaw } = useModuleI18n('features/config-metadata')
+const { translateIfKey, resolveConfigText } = useConfigTextResolver(props)
 
 const filteredIterable = computed(() => {
   if (!props.iterable) return {}
